+++ src/lib.rs
@@ -2,7 +2,6 @@
 
 mod io;
 mod log;
-mod term_idx;
 mod timeout;
 
 fn start() {}
